新增用户信息缓存机制并优化相关业务逻辑
- 在缓存中存储用户和团队信息，减少数据库查询
- 优化用户与

+++ cmd/server/palace/internal/service/builder/pagination.go
@@ -29,8 +29,8 @@ func (p *paginationModuleBuilder) ToAPI(pagination types.Pagination) *api.Pagina
 		return nil
 	}
 	return &api.PaginationReply{
-		PageNum:  int32(pagination.GetPageNum()),
-		PageSize: int32(pagination.GetPageSize()),
-		Total:    int64(pagination.GetTotal()),
+		PageNum:  pagination.GetPageNum(),
+		PageSize: pagination.GetPageSize(),
+		Total:    pagination.GetTotal(),
 	}
 }
